+++ src/lib/components/Carousel/Carousel.styles.ts
