+++ src/modules/main/components/ImportantContainer.tsx
@@ -4,7 +4,7 @@ import BorderBottom from '../../../lib/components/Border';
 import Carousel from '../../../lib/components/Carousel/Carousel';
 import { PostCard } from '../../../lib/components/Posts/Cards/PostCard';
 import { RootStateType } from '../../../store/rootReducer';
-import { useStyles } from '../styles/ImportantContainer.style';
+import { useStyles } from '../styles/ImportantContainer.styles';
 import {
   fetchImportantPosts,
   setImportantLoadingStatus,
